Fix decode/endpoint bugs found verifying routes against dahlia
Multi-resource verification sweep against the pinned 2026-06-24.dahlia
schemas surfaced nine correctness defects (silent data loss, decode
crashes, or a broken endpoint). All build clean and tests pass.

Decode-to-nil typos (property name didn't match the .convertFromSnakeCase
key, so the field never decoded):
- PaymentIntent.onBehalfOn -> onBehalfOf (on_behalf_of)
- Customer.preferredLocals -> preferredLocales (preferred_locales)
- Person.registeredAddres -> registeredAddress (registered_address)
- Invoice.markedUncollectableAt -> markedUncollectibleAt (marked_uncollectible_at)

Decode crashes:
- Coupon.percentOff and SubscriptionTransferData.amountPercent were Int
  but the API fields are floats; fractional values threw. Now Decimal.
- InvoiceBillingReason lacked automatic_pending_invoice_item_invoice and
  quote_accept; such invoices threw (no fallback case). Cases added.

Endpoint / request bugs:
- TransferReversalRoutes.retrieve/update built transfers/{id}/reversals/{id},
  using the reversal id as the transfer id (404). Now transfers/{transfer}/reversals/{id}.
- PromotionCodesRoutes.create sent body["expiresAt"]; the API param is
  expires_at, so the value was silently dropped. Fixed.

+++ Sources/StripeKit/Billing/Subscriptions/Subscription.swift
@@ -483,11 +483,11 @@ public enum SubscriptionTrialSettingsEndBehaviorMissingPaymentMethod: String, Co
 
 public struct SubscriptionTransferData: Codable, Sendable {
     /// A non-negative decimal between 0 and 100, with at most two decimal places. This represents the percentage of the subscription invoice subtotal that will be transferred to the destination account. By default, the entire amount is transferred to the destination.
-    public var amountPercent: Int?
+    public var amountPercent: Decimal?
     /// The account where funds from the payment will be transferred to upon payment success.
     @Expandable<ConnectAccount> public var destination: String?
-    
-    public init(amountPercent: Int? = nil, destination: String? = nil) {
+
+    public init(amountPercent: Decimal? = nil, destination: String? = nil) {
         self.amountPercent = amountPercent
         self._destination = Expandable(id: destination)
     }
